src/agents/chart_generator/chart_generator/requirements.txt
pandas
matplotlib
google-generativeai